Renamed edmundoptics to eo, for simplicity
